Worker Deployment API (#783)
* Modified get_instance_type() to include categorization of instances

* fetches various instance categories and corresponding instance types for the user

* Added Acclerated computing instances to instance types

* Fix /workers and /workers/<int:id> routes

* Added workers delete route

* Refactored Terraform and Provider classes

* Backward compatiblity bug fix

* Added states in workers deployment

* moved worker-api to domain folder

* Refactor AWS_Serverfull class to support worker deployment

* Read worker config from domain env variables

* write env vars after deploying domain

* Use domain credential to deploy worker

* Support pre-built loadings of plugins on terraform init

* Update association requests response

* Added general purpose instances to cli instance list

* removed domain/apps/core/infra

* Remove infra files from app/domain/src

* Add test route to check domain deployment

* fix: change pysyft branch to dev (#784)

* Minor fixes

* Added suggested changes

* Remove dummy infra route

* format files with black

+++ apps/infrastructure/providers/aws/aws.py
@@ -18,23 +18,16 @@ class AWS(Provider):
         os.makedirs(credentials_dir, exist_ok=True)
         self.cred_file = os.path.join(credentials_dir, "credentials.json")
 
-        with open(self.cred_file, "w") as f:
-            json.dump(vars(config.credentials.cloud), f, indent=2, sort_keys=False)
-
-        self.region = config.vpc.region
-        self.av_zones = config.vpc.av_zones
+        if not os.path.exists(self.cred_file):
+            with open(self.cred_file, "w") as f:
+                json.dump(vars(config.credentials.cloud), f, indent=2, sort_keys=False)
 
         self.tfscript += terrascript.provider.aws(
-            region=self.region, shared_credentials_file=self.cred_file
+            region=self.config.vpc.region, shared_credentials_file=self.cred_file
         )
 
-        # Build the Infrastructure
         self.vpc = None
         self.subnets = []
-        self.build_vpc()
-        self.build_igw()
-        self.build_public_rt()
-        self.build_subnets()
 
     def build_vpc(self):
         """Adds a VPC."""
@@ -92,7 +85,7 @@ class AWS(Provider):
         - one Route table : Routes the traffic from the NAT gateway to the private subnet
         """
 
-        for i, av_zone in enumerate(self.av_zones):
+        for i, av_zone in enumerate(self.config.vpc.av_zones):
             private_subnet = resource.aws_subnet(
                 f"private-subnet-{i}",
                 vpc_id=var(self.vpc.id),
